FIX: added a cast to int so that the priority value (which is a short) is correctly serialized as a string
GitOrigin-RevId: 0206ccd600af09962f880015406cbd3c5b7459d9

+++ dtests/node_data/clients/cpp/src/api/qpid/common/formatter/QpidDecoder.cpp
@@ -46,7 +46,8 @@ void QpidDecoder::write(Writer *writer,HeaderProperty property, BooleanReader re
 void QpidDecoder::write(Writer *writer,HeaderProperty property, ShortIntReader reader) const {
     uint8_t value = (message.*reader)();
     
-    writer->write(KeyValue(property.name, super::decodeValue(value)));
+    writer->write(KeyValue(property.name, 
+	    super::decodeValue(static_cast<int>(value))));
 }
